requirements.txt
Flask==2.3.2
tensorflow==2.18.0
numpy==2.0.2
Pillow==10.3.0
Werkzeug==3.0.6
Flask-SQLAlchemy
Flask-Login
flask-cors
pyjwt
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
pyOpenSSL